Endpoint para agregar un amigo

+++ src/main/java/dtos/TradeCreateDTO.java
@@ -4,7 +4,7 @@ package dtos;
  * Created by Palumbo on 29/09/2014.
  */
 public class TradeCreateDTO {
-    public int friendId;
-    public int friendItemId;
-    public int userItemId;
+    public long friendId;
+    public long friendItemId;
+    public long userItemId;
 }
